Don't populate the DB from MAL on failed searches
This removes the feature that was populating the database with "thin"
objects imported from MAL when a user was searching for a work that
did not exist in our database, since this feature was causing a lot of
duplicates.

It also uses the recently introduced `Category` table in order to
unify the handling in the `Manga` and `Anime` cases.