Remove duplicate ratings (#158)
Currently, we can have several ratings for a given (user, work) pair -
which is not good. The current code is not supposed to let that happen;
however this happened in practice in the past and the database is
cluttered with a few of those. This patch changes it so that:

 - There is now a UNIQUE index on the (user, work) pair of a rating
 - A SQL query (run in the migration) ensures that only the latter of
   several ratings is kept in case there was previously duplicates for
   the same (user, work) pair before creation of the UNIQUE index

Fixes #123.