Go back to using a static dictionnary for Work categories
This patch changes the way that categories are stored in the code.
Previously, we were using a specific table in the database, which had
disadvantages (outlined in #157). This moves that information back to
code, and is implemented using a custom Django field that allows to keep
the ease-of-use of writing statements such as:

  work = Work.objects.filter(category='anime').last()
  work.category.slug # 'anime'
  str(work.category) # 'Anime'

even though the `Work.category` field is ultimately a mere integer
field.

Fixes #157