Use init file for settings
Previously, we were using a mess of various settings files for Mangaki
settings.  Following the goal to separate configuration from code, we
are now using a separate .ini file that contains the configuration and
is parsed by Mangaki's settings.py module.

One of the reasons for this is that it should simplify the deployment of
new Mangaki instances.