Make Vagrant use a shell provisionner
Previously Vagrant was using Ansible, but the provisionement was
broken. Once we use Ansible for production, we should also use Ansible
for the Vagrant development environment, but in the meantime, let's
use a simple shell script instead.

This also removes the `CREATE EXTENSIONS` from the migrations since
they must be run as superuser (which the `django` user is not supposed
to be). They are outlined in the `README.md` file, and done on setup
by the shell provisioner instead.