README: simplify setup, use PostgreSQL local auth by default (#656)
With PostgreSQL, it is unnecessary to define database passwords for a
local development environment. We can simply use the "local" auth of
PostgreSQL, which allows UNIX users to connect without password to the
PG users of the same name. This greatly improves the ease of setup, and
removes the need for a config.sh.